feat: Add 1U formula to Type 1U color cipher algorithm
This commit adds the 1U formula to the Type 1U color cipher algorithm in the `method.md` file. The formula is used to distribute a list of objects along a line based on their specified frequency. This enhancement improves the mapping of values to colors based on the frequency of the characters, ensuring that frequently occurring characters have more colors assigned to them.